Made 'platform_getVolumeLevel' and 'platform_getMicLevel' API functions return a normalised 8 bit value ranging from 0 to 255 (TG-293 #closed)

+++ openrtx/include/interfaces/platform.h
@@ -58,11 +58,11 @@ typedef struct
 {
     char name[10];          /* Manufacturer-assigned hardware name.             */
 
-    uint16_t uhf_maxFreq;   /* Upper bound for UHF band, in MHz.               */
-    uint16_t uhf_minFreq;   /* Lower bound for UHF band, in MHz.               */
+    uint16_t uhf_maxFreq;   /* Upper bound for UHF band, in MHz.                */
+    uint16_t uhf_minFreq;   /* Lower bound for UHF band, in MHz.                */
 
-    uint16_t vhf_maxFreq;   /* Upper bound for VHF band, in MHz.               */
-    uint16_t vhf_minFreq;   /* Lower bound for VHF band, in MHz.               */
+    uint16_t vhf_maxFreq;   /* Upper bound for VHF band, in MHz.                */
+    uint16_t vhf_minFreq;   /* Lower bound for VHF band, in MHz.                */
 
     uint8_t  _unused  : 4,
              uhf_band : 1,  /* Device allows UHF band operation.                */
@@ -89,14 +89,16 @@ void platform_terminate();
 float platform_getVbat();
 
 /**
- * This function reads and returns the current microphone input level.
+ * This function reads and returns the current microphone input level as a
+ * normalised value between 0 and 255.
  */
-float platform_getMicLevel();
+uint8_t platform_getMicLevel();
 
 /**
- * This function reads and returns the current volume selector level.
+ * This function reads and returns the current volume selector level as a
+ * normalised value between 0 and 255.
  */
-float platform_getVolumeLevel();
+uint8_t platform_getVolumeLevel();
 
 /**
  * This function reads and returns the current channel selector level.
